+++ explcheck/testfiles/e427-01.lua
@@ -12,7 +12,7 @@ local issues, results = state.issues, state.results
 assert(#issues.errors == 1)
 assert(#issues.warnings == 0)
 
-local expected_line_numbers = {{4, 6}}
+local expected_line_numbers = {{6, 6}}
 for index, err in ipairs(sort_issues(issues.errors)) do
   assert(err[1] == "e427")
   assert(err[2] == "comparison conditional without signature `:nnTF`")